backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Pillow==10.1.0
img2pdf==0.5.1
Werkzeug==3.0.1

